(Feat) 저장 슬롯 기능 임시
아직 기능은 완벽히 제대로 완성되지 않음
메인 메뉴 DOTween 컴포넌트 리팩토링 필요. 스크립트로 관리를 해야할 것 같음

+++ Assets/Scripts/Data Scripts/StateUpgradeData.cs
@@ -10,8 +10,8 @@ public class StateUpgradeData : ScriptableObject
     [PreviewField(50, ObjectFieldAlignment.Left)]
     public Sprite icon;
     public string stateName;
-    public byte level;
-    public byte maxLevel;
+    public short level;
+    public short maxLevel;
     public short totalIncreaseValue;
     public short totalMoreValue;
     [InfoBox("n% 스탯 증가\nState * Increase * More")]
@@ -28,7 +28,7 @@ public class StateUpgradeData : ScriptableObject
     // 행운
     // 시작 레벨
 
-    public void ApplyState(byte level)
+    public void ApplyState(short level)
     {
         // 혹시나 스탯 값이 업데이트로 변경될 때를 대비해
         // 레벨을 기준으로 다시 적용시키는 방식
